+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 <!-- ### Security 🛡️-->
 
 ## [Unreleased]
+* **Ontimize SQLStatementHandler**: Allow searching for a pattern in columns with any data type
 * **Keycloak**: Allow to use the Keycloak authentication on non-http calls
 ## [5.9.0] - 2023-07-31
 ### Added ✔️